requirements.txt
django-generic-m2m
South
django-model-utils
